updated auth clarity
Updated navbar to reflect whether user is signed in or not. Made navbar login a nextAuth hook instead of a redirect. created consistent response returns (each response typically has a message and a status). Reverted nextAuth session strategy to jwt as nextAuth database session appears to not support credentials provider.